backend/requirements.txt
Flask
flask-mysqldb
flask-cors
Pillow
pytesseract
 